requirements.txt
alembic>=1.13.1
asyncpg>=0.29.0
beautifulsoup4>=4.12.3
celery>=5.3.6
fastapi>=0.111.0
google-generativeai>=0.4.1
httpx>=0.27.0
neo4j>=5.19.0
numpy>=1.26.4
openpyxl>=3.1.2
pandas>=2.2.1
passlib[bcrypt]>=1.7.4
playwright>=1.44.0
pydantic-settings>=2.2.1
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
redis>=5.0.3
reportlab>=4.1.0
scikit-learn>=1.4.1.post1
sqlalchemy[asyncio]>=2.0.29
structlog>=24.1.0
uvicorn[standard]>=0.29.0
-e ./packages/shared-config
-e ./packages/shared-types
-e ./packages/shared-utils
